- split WinePrefixContainerPanel in multiple classes
