+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/containers/ContainerPanelFactory.java
@@ -19,6 +19,7 @@
 package org.phoenicis.javafx.views.mainwindow.containers;
 
 import org.phoenicis.containers.dto.ContainerDTO;
+import org.phoenicis.containers.wine.WinePrefixContainerController;
 import org.phoenicis.engines.dto.EngineVersionDTO;
 import org.phoenicis.javafx.views.common.ThemeManager;
 
@@ -34,10 +35,10 @@ public class ContainerPanelFactory<T extends AbstractContainerPanel<C>, C extend
         this.entityClazz = entityClazz;
     }
 
-    public T createContainerPanel(C containerDTO, ThemeManager themeManager, List<EngineVersionDTO> engineVersions) {
+    public T createContainerPanel(C containerDTO, ThemeManager themeManager, List<EngineVersionDTO> engineVersions, WinePrefixContainerController winePrefixContainerController) {
         try {
-            return this.clazz.getConstructor(entityClazz, ThemeManager.class, List.class).newInstance(containerDTO,
-                    themeManager, engineVersions);
+            return this.clazz.getConstructor(entityClazz, ThemeManager.class, List.class, WinePrefixContainerController.class).newInstance(containerDTO,
+                    themeManager, engineVersions, winePrefixContainerController);
         } catch (InstantiationException | IllegalAccessException | InvocationTargetException
                 | NoSuchMethodException e) {
             throw new IllegalStateException(e);
